requirements.txt
apache-airflow==2.9.0
requests==2.31.0
sqlalchemy==2.0.25
pandas==2.2.1
psycopg2-binary==2.9.9
boto3==1.34.34